(fix) firebase-tools are now available to use for deploy (#215)
This commit will make firebase-tools available for use in deploy. The build was failing due to what seems to be a known issue: https://github.com/travis-ci/travis-ci/issues/8637. This commit will also only run CI on pull requests and merges to master, whereas before it either ran two builds on every commit or didn't run on merges to master.